Move region combo box to toolkit lib
