+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/appservice/AppServiceInfoAdvancedPanel.form
@@ -56,7 +56,7 @@
           <grid row="5" column="1" row-span="1" col-span="2" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
       </component>
-      <component id="312a0" class="com.microsoft.azure.toolkit.intellij.appservice.region.RegionComboBox" binding="selectorRegion">
+      <component id="312a0" class="com.microsoft.azure.toolkit.intellij.common.component.RegionComboBox" binding="selectorRegion">
         <constraints>
           <grid row="6" column="1" row-span="1" col-span="2" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
