+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/arm/CreateDeploymentForm.form
@@ -122,7 +122,7 @@
               <text value="Region:"/>
             </properties>
           </component>
-          <component id="e5995" class="com.microsoft.azure.toolkit.intellij.appservice.region.RegionComboBox" binding="regionCb" custom-create="true">
+          <component id="e5995" class="com.microsoft.azure.toolkit.intellij.common.component.RegionComboBox" binding="regionCb" custom-create="true">
             <constraints>
               <grid row="6" column="2" row-span="1" col-span="2" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
             </constraints>
