+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/redis/CreateRedisCacheForm.form
@@ -75,7 +75,7 @@
             </constraints>
             <properties/>
           </component>
-          <component id="37ea6" class="com.microsoft.azure.toolkit.intellij.appservice.region.RegionComboBox" binding="cbLocations" custom-create="true">
+          <component id="37ea6" class="com.microsoft.azure.toolkit.intellij.common.component.RegionComboBox" binding="cbLocations" custom-create="true">
             <constraints>
               <grid row="10" column="0" row-span="1" col-span="9" vsize-policy="0" hsize-policy="3" anchor="8" fill="1" indent="0" use-parent-layout="false">
                 <minimum-size width="330" height="-1"/>
