+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/storage/component/BaseStorageAccountCreationDialog.form
@@ -78,7 +78,7 @@
           <text value="Location:"/>
         </properties>
       </component>
-      <component id="400c" class="com.microsoft.azure.toolkit.intellij.appservice.region.RegionComboBox" binding="regionComboBox">
+      <component id="400c" class="com.microsoft.azure.toolkit.intellij.common.component.RegionComboBox" binding="regionComboBox">
         <constraints>
           <grid row="5" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="1" use-parent-layout="false"/>
         </constraints>
