+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/storage/component/BaseStorageAccountCreationDialog.java
@@ -6,7 +6,7 @@
 package com.microsoft.azure.toolkit.intellij.storage.component;
 
 import com.intellij.openapi.project.Project;
-import com.microsoft.azure.toolkit.intellij.appservice.region.RegionComboBox;
+import com.microsoft.azure.toolkit.intellij.common.component.RegionComboBox;
 import com.microsoft.azure.toolkit.intellij.common.component.resourcegroup.ResourceGroupComboBox;
 import com.microsoft.azure.toolkit.intellij.common.AzureDialog;
 import com.microsoft.azure.toolkit.intellij.common.component.SubscriptionComboBox;
