+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/intellij/ui/ApplicationInsightsNewDialog.form
@@ -59,7 +59,7 @@
         </constraints>
         <properties/>
       </component>
-      <component id="f8ade" class="com.microsoft.azure.toolkit.intellij.appservice.region.RegionComboBox" binding="comboReg" custom-create="true">
+      <component id="f8ade" class="com.microsoft.azure.toolkit.intellij.common.component.RegionComboBox" binding="comboReg" custom-create="true">
         <constraints>
           <grid row="4" column="1" row-span="1" col-span="2" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
